Performance and style improvements
I added a prefetch method for all the photos. It removes 2-3 seconds from the query. Of course, the other performance issue is generating a 3.5MB json file. Ideally, this is broken into separate queries, but for now the cache works great.
I've also added a config setting to disable the cache.
Improved admin layout, including adding thumbnails for the images. Ideally, these could be sortable, so that SC team would be able to see which stairs have images.
Adjusted the styles and navigation as well.